Finally got a little something working with the salesforce api.
The Auth is working now using the username and password auth flow.

From reading the salesforce rest api, in order to make new orders:
- create all the order items
- have a list of order items to be attached to the order
- have an account to which the order will be attached

So in preparation for this, I added a whole bunch more angular files to control all the different pages
and widgets that will be needed for creating and managing these different types of salesforce objects.